refactor(TemplateEditPage): TemplateEdit 및 useTemplateEdit 삭제
