+++ frontend/src/pages/TemplateEditPage/TemplateEditPage.tsx
@@ -1,7 +1,16 @@
-import type { Template } from '@/types';
+import { PlusIcon } from '@/assets/images';
+import { Button, Input, SelectList, SourceCodeEditor, Text, CategoryDropdown, TagInput } from '@/components';
+import { useInput } from '@/hooks';
+import { useCategory } from '@/hooks/category';
+import { useSourceCodeSelectList } from '@/hooks/template';
+import { useToast } from '@/hooks/useToast';
+import { useTemplateEditMutation } from '@/queries/templates';
+import { theme } from '@/style/theme';
+import type { Template, TemplateEditRequest } from '@/types';
 
-import { TemplateEdit } from './components';
-import { useTemplateEdit } from './hooks';
+import { useSourceCode } from '../../hooks/template/useSourceCode';
+import { useTag } from '../../hooks/template/useTag';
+import * as S from './TemplateEditPage.style';
 
 interface Props {
   template: Template;
@@ -9,9 +18,148 @@ interface Props {
 }
 
 const TemplateEditPage = ({ template, toggleEditButton }: Props) => {
-  const props = useTemplateEdit({ template, toggleEditButton });
+  const categoryProps = useCategory(template.category);
 
-  return <TemplateEdit {...props} />;
+  const [title, handleTitleChange] = useInput(template.title);
+  const [description, handleDescriptionChange] = useInput(template.description);
+
+  const {
+    sourceCodes,
+    deleteSourceCodeIds,
+    handleFilenameChange,
+    handleCodeChange,
+    addNewEmptySourceCode,
+    handleDeleteSourceCode,
+  } = useSourceCode(template.sourceCodes);
+
+  const initTags = template.tags.map((tag) => tag.name);
+  const tagProps = useTag(initTags);
+
+  const { currentFile, sourceCodeRefs, handleSelectOption } = useSourceCodeSelectList(sourceCodes);
+
+  const { mutateAsync: updateTemplate, error } = useTemplateEditMutation(template.id);
+
+  const { failAlert } = useToast();
+
+  const validateTemplate = () => {
+    if (!title) {
+      return '제목을 입력해주세요';
+    }
+
+    if (sourceCodes.filter((sourceCode) => !sourceCode.filename).length) {
+      return '파일명을 입력해주세요';
+    }
+
+    if (sourceCodes.filter((sourceCode) => !sourceCode.content).length) {
+      return '소스코드 내용을 입력해주세요';
+    }
+
+    return '';
+  };
+
+  const handleCancelButton = () => {
+    toggleEditButton();
+  };
+
+  const handleSaveButtonClick = async () => {
+    if (validateTemplate()) {
+      failAlert(validateTemplate());
+
+      return;
+    }
+
+    const orderedSourceCodes = sourceCodes.map((sourceCode, index) => ({
+      ...sourceCode,
+      ordinal: index + 1,
+    }));
+    const createSourceCodes = orderedSourceCodes.filter((sourceCode) => !sourceCode.id);
+    const updateSourceCodes = orderedSourceCodes.filter((sourceCode) => sourceCode.id);
+
+    const templateUpdate: TemplateEditRequest = {
+      title,
+      description,
+      createSourceCodes,
+      updateSourceCodes,
+      deleteSourceCodeIds,
+      categoryId: categoryProps.currentValue.id,
+      tags: tagProps.tags,
+    };
+
+    try {
+      await updateTemplate({ id: template.id, template: templateUpdate });
+      toggleEditButton();
+    } catch (error) {
+      console.error('Failed to update template:', error);
+    }
+  };
+
+  return (
+    <S.TemplateEditContainer>
+      <S.MainContainer>
+        <CategoryDropdown {...categoryProps} />
+
+        <S.UnderlineInputWrapper>
+          <Input size='xlarge' variant='text'>
+            <Input.TextField placeholder='제목을 입력해주세요' value={title} onChange={handleTitleChange} />
+          </Input>
+        </S.UnderlineInputWrapper>
+
+        <Input size='large' variant='text'>
+          <Input.TextField
+            placeholder='이 템플릿을 언제 다시 쓸 것 같나요?'
+            value={description}
+            onChange={handleDescriptionChange}
+          />
+        </Input>
+
+        {sourceCodes.map((sourceCode, index) => (
+          <SourceCodeEditor
+            key={index}
+            sourceCodeRef={(el) => (sourceCodeRefs.current[index] = el)}
+            filename={sourceCode.filename}
+            content={sourceCode.content}
+            onChangeContent={(newContent) => handleCodeChange(newContent, index)}
+            onChangeFilename={(newFilename) => handleFilenameChange(newFilename, index)}
+            handleDeleteSourceCode={() => handleDeleteSourceCode(index)}
+            filenameAutoFocus={index !== 0}
+          />
+        ))}
+
+        <Button
+          size='medium'
+          variant='contained'
+          buttonColor={theme.color.light.primary_50}
+          fullWidth
+          onClick={addNewEmptySourceCode}
+        >
+          <PlusIcon width={14} height={14} aria-label='소스코드 추가' />
+        </Button>
+
+        <TagInput {...tagProps} />
+
+        <S.ButtonGroup>
+          <Button size='medium' variant='outlined' onClick={handleCancelButton}>
+            취소
+          </Button>
+          <Button size='medium' variant='contained' onClick={handleSaveButtonClick} disabled={sourceCodes.length === 0}>
+            저장
+          </Button>
+        </S.ButtonGroup>
+
+        {error && <Text.Medium color={theme.color.light.analogous_primary_400}>Error: {error.message}</Text.Medium>}
+      </S.MainContainer>
+
+      <S.SidebarContainer>
+        <SelectList>
+          {sourceCodes.map((sourceCode, index) => (
+            <SelectList.Option key={index} onClick={handleSelectOption(index)} isSelected={currentFile === index}>
+              {sourceCode.filename}
+            </SelectList.Option>
+          ))}
+        </SelectList>
+      </S.SidebarContainer>
+    </S.TemplateEditContainer>
+  );
 };
 
 export default TemplateEditPage;
