--- frontend/src/pages/TemplateEditPage/components/index.ts
@@ -1,1 +1,0 @@
-export { default as TemplateEdit } from './TemplateEdit/TemplateEdit';
