--- frontend/src/pages/TemplateEditPage/hooks/index.ts
@@ -1,1 +1,0 @@
-export { useTemplateEdit } from './useTemplateEdit';
